updated the code base: moved the ./routes/socket folder to ./socket.  Updated the /routes/game folder and finalized the /allGames endpoint to send the ID and playercount of all active games.
